Actually use the correct type for entity ids in the event broker

+++ src/main/raptor/engine/event/EventBroker.java
@@ -9,18 +9,18 @@ import java.util.Queue;
 
 public class EventBroker implements IEventBroker {
 	private final Queue<IEvent> eventQueue;
-	private final Map<Integer, IEventDestination> entityRegistration;
-	private final Map<String, List<Integer>> eventRegistration;
+	private final Map<Long, IEventDestination> entityRegistration;
+	private final Map<String, List<Long>> eventRegistration;
 
-	private final Queue<Integer> triggerEntityIds;
+	private final Queue<Long> triggerEntityIds;
 	private final Queue<IEvent> triggerEvents;
 
 	public EventBroker() {
 		this.eventQueue = new LinkedList<IEvent>();
-		this.entityRegistration = new HashMap<Integer, IEventDestination>();
-		this.eventRegistration = new HashMap<String, List<Integer>>();
+		this.entityRegistration = new HashMap<Long, IEventDestination>();
+		this.eventRegistration = new HashMap<String, List<Long>>();
 
-		this.triggerEntityIds = new LinkedList<Integer>();
+		this.triggerEntityIds = new LinkedList<Long>();
 		this.triggerEvents = new LinkedList<IEvent>();
 	}
 
@@ -30,7 +30,7 @@ public class EventBroker implements IEventBroker {
 	}
 
 	@Override
-	public void trigger(final int entityId, final IEvent event) {
+	public void trigger(final long entityId, final IEvent event) {
 		if (!entityRegistration.containsKey(entityId))
 			return;
 
@@ -42,15 +42,15 @@ public class EventBroker implements IEventBroker {
 	public void distribute() {
 		while(!eventQueue.isEmpty()) {
 			final IEvent event = eventQueue.poll();
-			final List<Integer> registrees = eventRegistration.get(event.getName());
+			final List<Long> registrees = eventRegistration.get(event.getName());
 			if (registrees == null || registrees.size() <= 0)
 				continue;
-			for (final Integer id : registrees)
+			for (final Long id : registrees)
 				entityRegistration.get(id).send(event);
 		}
 
 		while(!triggerEntityIds.isEmpty()) {
-			final int triggerEntityId = triggerEntityIds.poll();
+			final long triggerEntityId = triggerEntityIds.poll();
 			final IEvent triggerEvent = triggerEvents.poll();
 			if (!entityRegistration.containsKey(triggerEntityId))
 				continue;
@@ -59,48 +59,48 @@ public class EventBroker implements IEventBroker {
 	}
 
 	@Override
-	public IEventSource register(final int entityId) {
+	public IEventSource register(final long entityId) {
 		final EventExchange exchange = new EventExchange();
 		entityRegistration.put(entityId, exchange);
 		return exchange;
 	}
 
 	@Override
-	public void register(final int entityId, final String eventName) {
+	public void register(final long entityId, final String eventName) {
 		if (!isRegistered(entityId))
 			register(entityId);
 		if (!brokersEvent(eventName))
-			eventRegistration.put(eventName, new ArrayList<Integer>());
-		final List<Integer> registrees = eventRegistration.get(eventName);
+			eventRegistration.put(eventName, new ArrayList<Long>());
+		final List<Long> registrees = eventRegistration.get(eventName);
 		if (!registrees.contains(entityId))
 			registrees.add(entityId);
 	}
 
 	@Override
-	public void unregister(final int entityId) {
-		for (final List<Integer> registrees : eventRegistration.values())
+	public void unregister(final long entityId) {
+		for (final List<Long> registrees : eventRegistration.values())
 			registrees.remove(entityId);
 		entityRegistration.remove(entityId);
 	}
 
 	@Override
-	public void unregister(final int entityId, final String eventName) {
+	public void unregister(final long entityId, final String eventName) {
 		if (!brokersEvent(eventName))
 			return;
-		final List<Integer> registrees = eventRegistration.get(eventName);
+		final List<Long> registrees = eventRegistration.get(eventName);
 		registrees.remove(entityId);
 	}
 
 	@Override
-	public boolean isRegistered(final int entityId) {
+	public boolean isRegistered(final long entityId) {
 		return entityRegistration.containsKey(entityId);
 	}
 
 	@Override
-	public boolean isRegistered(final int entityId, final String eventName) {
+	public boolean isRegistered(final long entityId, final String eventName) {
 		if (!isRegistered(entityId) || !brokersEvent(eventName))
 			return false;
-		final List<Integer> registrees = eventRegistration.get(eventName);
+		final List<Long> registrees = eventRegistration.get(eventName);
 		return registrees.contains(entityId);
 	}
